add durable statestore example
